SHIM: Down-convert 0.4 vdW
