+++ openff/interchange/smirnoff/_create.py
@@ -231,6 +231,11 @@ def _vdw(interchange: Interchange, force_field: ForceField, topology: Topology):
     if "vdW" not in force_field.registered_parameter_handlers:
         return
 
+    if force_field["vdW"].version == Version("0.4"):
+        from openff.interchange.smirnoff._nonbonded import _downconvert_vdwhandler
+
+        _downconvert_vdwhandler(force_field["vdW"])
+
     interchange.collections.update(
         {
             "vdW": SMIRNOFFvdWCollection.create(
